Refactor FDC3Error as static serialize/deserialize methods
